Extract useMotionPreference hook
